requirements.txt
langchain==0.3.23
langchain-community==0.3.21
openai==1.17.0
faiss-cpu==1.7.4
streamlit==1.32.2
python-dotenv==1.1.0
tqdm==4.66.1
unstructured==0.11.6

